Added version 1.4.589

+++ chromium/chrome/browser/ui/webui/signin/user_manager_screen_handler.cc
@@ -5,6 +5,7 @@
 #include "chrome/browser/ui/webui/signin/user_manager_screen_handler.h"
 
 #include <stddef.h>
+
 #include <utility>
 #include <vector>
 
@@ -45,9 +46,11 @@
 #include "chrome/common/url_constants.h"
 #include "chrome/grit/chromium_strings.h"
 #include "chrome/grit/generated_resources.h"
+#include "chrome/grit/google_chrome_strings.h"
 #include "components/prefs/pref_service.h"
 #include "components/proximity_auth/screenlock_bridge.h"
 #include "components/signin/core/account_id/account_id.h"
+#include "components/signin/core/common/profile_management_switches.h"
 #include "content/public/browser/notification_service.h"
 #include "content/public/browser/storage_partition.h"
 #include "content/public/browser/web_contents.h"
@@ -99,6 +102,8 @@ const char kJsApiUserManagerRemoveUserWarningLoadStats[] =
     "removeUserWarningLoadStats";
 const char kJsApiUserManagerGetRemoveWarningDialogMessage[] =
     "getRemoveWarningDialogMessage";
+const char kJsApiUserManagerAreAllProfilesLocked[] =
+    "areAllProfilesLocked";
 const size_t kAvatarIconSize = 180;
 const int kMaxOAuthRetries = 3;
 
@@ -135,12 +140,16 @@ extensions::ScreenlockPrivateEventRouter* GetScreenlockRouter(
 }
 
 bool IsGuestModeEnabled() {
+  if (switches::IsMaterialDesignUserManager())
+    return true;
   PrefService* service = g_browser_process->local_state();
   DCHECK(service);
   return service->GetBoolean(prefs::kBrowserGuestModeEnabled);
 }
 
 bool IsAddPersonEnabled() {
+  if (switches::IsMaterialDesignUserManager())
+    return true;
   PrefService* service = g_browser_process->local_state();
   DCHECK(service);
   return service->GetBoolean(prefs::kBrowserAddPersonEnabled);
@@ -295,7 +304,7 @@ UserManagerScreenHandler::~UserManagerScreenHandler() {
 
 void UserManagerScreenHandler::ShowBannerMessage(
     const base::string16& message) {
-  web_ui()->CallJavascriptFunction(
+  web_ui()->CallJavascriptFunctionUnsafe(
       "login.AccountPickerScreen.showBannerMessage",
       base::StringValue(message));
 }
@@ -308,14 +317,14 @@ void UserManagerScreenHandler::ShowUserPodCustomIcon(
       icon_options.ToDictionaryValue();
   if (!icon || icon->empty())
     return;
-  web_ui()->CallJavascriptFunction(
+  web_ui()->CallJavascriptFunctionUnsafe(
       "login.AccountPickerScreen.showUserPodCustomIcon",
       base::StringValue(account_id.GetUserEmail()), *icon);
 }
 
 void UserManagerScreenHandler::HideUserPodCustomIcon(
     const AccountId& account_id) {
-  web_ui()->CallJavascriptFunction(
+  web_ui()->CallJavascriptFunctionUnsafe(
       "login.AccountPickerScreen.hideUserPodCustomIcon",
       base::StringValue(account_id.GetUserEmail()));
 }
@@ -333,10 +342,10 @@ void UserManagerScreenHandler::SetAuthType(
     return;
 
   user_auth_type_map_[account_id.GetUserEmail()] = auth_type;
-  web_ui()->CallJavascriptFunction("login.AccountPickerScreen.setAuthType",
-                                   base::StringValue(account_id.GetUserEmail()),
-                                   base::FundamentalValue(auth_type),
-                                   base::StringValue(auth_value));
+  web_ui()->CallJavascriptFunctionUnsafe(
+      "login.AccountPickerScreen.setAuthType",
+      base::StringValue(account_id.GetUserEmail()),
+      base::FundamentalValue(auth_type), base::StringValue(auth_value));
 }
 
 proximity_auth::ScreenlockBridge::LockHandler::AuthType
@@ -372,7 +381,8 @@ void UserManagerScreenHandler::HandleInitialize(const base::ListValue* args) {
   args->GetString(0, &url_hash_);
 
   SendUserList();
-  web_ui()->CallJavascriptFunction("cr.ui.UserManager.showUserManagerScreen",
+  web_ui()->CallJavascriptFunctionUnsafe(
+      "cr.ui.UserManager.showUserManagerScreen",
       base::FundamentalValue(IsGuestModeEnabled()),
       base::FundamentalValue(IsAddPersonEnabled()));
 
@@ -446,7 +456,8 @@ void UserManagerScreenHandler::HandleAuthenticatedLaunchUser(
   // In order to support the upgrade case where we have a local hash but no
   // password token, the user perform a full online reauth.
   UserManager::ShowReauthDialog(web_ui()->GetWebContents()->GetBrowserContext(),
-                                email_address_);
+                                email_address_,
+                                signin_metrics::Reason::REASON_UNLOCK);
 }
 
 void UserManagerScreenHandler::HandleRemoveUser(const base::ListValue* args) {
@@ -465,6 +476,16 @@ void UserManagerScreenHandler::HandleRemoveUser(const base::ListValue* args) {
 
   DCHECK(profiles::IsMultipleProfilesEnabled());
 
+  if (switches::IsMaterialDesignUserManager() &&
+      profiles::AreAllProfilesLocked()) {
+    web_ui()->CallJavascriptFunctionUnsafe(
+        "cr.webUIListenerCallback",
+        base::StringValue("show-error-dialog"),
+        base::StringValue(l10n_util::GetStringUTF8(
+            IDS_USER_MANAGER_REMOVE_PROFILE_PROFILES_LOCKED_ERROR)));
+    return;
+  }
+
   // The callback is run if the only profile has been deleted, and a new
   // profile has been created to replace it.
   webui::DeleteProfileAtPath(profile_path,
@@ -482,6 +503,19 @@ void UserManagerScreenHandler::HandleLaunchGuest(const base::ListValue* args) {
     // guest mode.
     NOTREACHED();
   }
+}
+
+void UserManagerScreenHandler::HandleAreAllProfilesLocked(
+    const base::ListValue* args) {
+  std::string webui_callback_id;
+  CHECK_EQ(1U, args->GetSize());
+  bool success = args->GetString(0, &webui_callback_id);
+  DCHECK(success);
+
+  AllowJavascript();
+  ResolveJavascriptCallback(
+      base::StringValue(webui_callback_id),
+      base::FundamentalValue(profiles::AreAllProfilesLocked()));
 }
 
 void UserManagerScreenHandler::HandleLaunchUser(const base::ListValue* args) {
@@ -575,9 +609,9 @@ void UserManagerScreenHandler::HandleRemoveUserWarningLoadStats(
       stats_success &= item.success;
     }
     if (stats_success) {
-      web_ui()->CallJavascriptFunction("updateRemoveWarningDialog",
-                                       base::StringValue(profile_path.value()),
-                                       return_value);
+      web_ui()->CallJavascriptFunctionUnsafe(
+          "updateRemoveWarningDialog", base::StringValue(profile_path.value()),
+          return_value);
       return;
     }
   }
@@ -599,9 +633,9 @@ void UserManagerScreenHandler::RemoveUserDialogLoadStatsCallback(
     stat->SetBooleanWithoutPathExpansion("success", item.success);
     return_value.SetWithoutPathExpansion(item.category, std::move(stat));
   }
-  web_ui()->CallJavascriptFunction("updateRemoveWarningDialog",
-                                   base::StringValue(profile_path.value()),
-                                   return_value);
+  web_ui()->CallJavascriptFunctionUnsafe(
+      "updateRemoveWarningDialog", base::StringValue(profile_path.value()),
+      return_value);
 }
 
 void UserManagerScreenHandler::HandleGetRemoveWarningDialogMessage(
@@ -632,10 +666,9 @@ void UserManagerScreenHandler::HandleGetRemoveWarningDialogMessage(
   base::StringValue message = base::StringValue(
       l10n_util::GetPluralStringFUTF16(message_id, total_count));
 
-  web_ui()->CallJavascriptFunction("updateRemoveWarningDialogSetMessage",
-                                   base::StringValue(profile_path),
-                                   message,
-                                   base::FundamentalValue(total_count));
+  web_ui()->CallJavascriptFunctionUnsafe(
+      "updateRemoveWarningDialogSetMessage", base::StringValue(profile_path),
+      message, base::FundamentalValue(total_count));
 }
 
 void UserManagerScreenHandler::OnGetTokenInfoResponse(
@@ -649,7 +682,8 @@ void UserManagerScreenHandler::OnOAuthError() {
   DCHECK(!email_address_.empty());
   oauth_client_.reset();
   UserManager::ShowReauthDialog(web_ui()->GetWebContents()->GetBrowserContext(),
-                                email_address_);
+                                email_address_,
+                                signin_metrics::Reason::REASON_UNLOCK);
 }
 
 void UserManagerScreenHandler::OnNetworkError(int response_code) {
@@ -688,6 +722,10 @@ void UserManagerScreenHandler::RegisterMessages() {
   web_ui()->RegisterMessageCallback(
       kJsApiUserManagerGetRemoveWarningDialogMessage,
       base::Bind(&UserManagerScreenHandler::HandleGetRemoveWarningDialogMessage,
+                 base::Unretained(this)));
+  web_ui()->RegisterMessageCallback(
+      kJsApiUserManagerAreAllProfilesLocked,
+      base::Bind(&UserManagerScreenHandler::HandleAreAllProfilesLocked,
                  base::Unretained(this)));
 
   const content::WebUI::MessageCallback& kDoNothingCallback =
@@ -826,6 +864,15 @@ void UserManagerScreenHandler::GetLocalizedValues(
                                 base::string16());
   localized_strings->SetString("multiProfilesOwnerPrimaryOnlyMsg",
                                 base::string16());
+
+  // Error message when trying to add a profile while all profiles are locked.
+  localized_strings->SetString("addProfileAllProfilesLockedError",
+      l10n_util::GetStringUTF16(
+          IDS_USER_MANAGER_ADD_PROFILE_PROFILES_LOCKED_ERROR));
+  // Error message when trying to browse as guest while all profiles are locked.
+  localized_strings->SetString("browseAsGuestAllProfilesLockedError",
+      l10n_util::GetStringUTF16(
+          IDS_USER_MANAGER_GO_GUEST_PROFILES_LOCKED_ERROR));
 }
 
 void UserManagerScreenHandler::SendUserList() {
@@ -847,7 +894,8 @@ void UserManagerScreenHandler::SendUserList() {
     if (entry->IsOmitted())
       continue;
 
-    base::DictionaryValue* profile_value = new base::DictionaryValue();
+    std::unique_ptr<base::DictionaryValue> profile_value(
+        new base::DictionaryValue());
     base::FilePath profile_path = entry->GetPath();
 
     profile_value->SetString(kKeyUsername, entry->GetUserName());
@@ -888,11 +936,12 @@ void UserManagerScreenHandler::SendUserList() {
         g_browser_process->profile_manager()->GetProfileByPath(profile_path);
     profile_value->SetBoolean(kKeyIsProfileLoaded, profile != nullptr);
 
-    users_list.Append(profile_value);
+    users_list.Append(std::move(profile_value));
   }
 
-  web_ui()->CallJavascriptFunction("login.AccountPickerScreen.loadUsers",
-      users_list, base::FundamentalValue(IsGuestModeEnabled()));
+  web_ui()->CallJavascriptFunctionUnsafe(
+      "login.AccountPickerScreen.loadUsers", users_list,
+      base::FundamentalValue(IsGuestModeEnabled()));
 
   // This is the latest C++ code we have in the flow to show the UserManager.
   // This may be invoked more than once per UserManager lifetime; the
@@ -913,15 +962,13 @@ void UserManagerScreenHandler::ReportAuthenticationResult(
                    weak_ptr_factory_.GetWeakPtr()),
         ProfileMetrics::SWITCH_PROFILE_UNLOCK);
   } else {
-    web_ui()->CallJavascriptFunction(
-        "cr.ui.UserManager.showSignInError",
-        base::FundamentalValue(0),
+    web_ui()->CallJavascriptFunctionUnsafe(
+        "cr.ui.UserManager.showSignInError", base::FundamentalValue(0),
         base::StringValue(l10n_util::GetStringUTF8(
-            auth == ProfileMetrics::AUTH_FAILED_OFFLINE ?
-                IDS_LOGIN_ERROR_AUTHENTICATING_OFFLINE :
-                IDS_LOGIN_ERROR_AUTHENTICATING)),
-        base::StringValue(""),
-        base::FundamentalValue(0));
+            auth == ProfileMetrics::AUTH_FAILED_OFFLINE
+                ? IDS_LOGIN_ERROR_AUTHENTICATING_OFFLINE
+                : IDS_LOGIN_ERROR_AUTHENTICATING)),
+        base::StringValue(""), base::FundamentalValue(0));
   }
 }
 
@@ -955,17 +1002,12 @@ void UserManagerScreenHandler::Observe(
     int type,
     const content::NotificationSource& source,
     const content::NotificationDetails& details) {
-  switch (type) {
-    case chrome::NOTIFICATION_BROWSER_WINDOW_READY:
-      // Only respond to one Browser Window Ready event.
-      registrar_.Remove(this,
-                        chrome::NOTIFICATION_BROWSER_WINDOW_READY,
-                        content::NotificationService::AllSources());
-      OnBrowserWindowReady(content::Source<Browser>(source).ptr());
-    break;
-    default:
-      NOTREACHED();
-  }
+  DCHECK_EQ(chrome::NOTIFICATION_BROWSER_WINDOW_READY, type);
+
+  // Only respond to one Browser Window Ready event.
+  registrar_.Remove(this, chrome::NOTIFICATION_BROWSER_WINDOW_READY,
+                    content::NotificationService::AllSources());
+  OnBrowserWindowReady(content::Source<Browser>(source).ptr());
 }
 
 // This callback is run after switching to a new profile has finished. This
